RE-144 fix: insight count long 타입 변경

+++ src/main/java/info/reinput/reinputworkspaceservice/folder/infra/client/ContentClient.java
@@ -23,7 +23,7 @@ public interface ContentClient {
     );
 
     @GetMapping("/insight/copy/{folderId}")
-    ApiResponse<Integer> copyInsight(
+    ApiResponse<Long> copyInsight(
             @PathVariable final Long folderId,
             @RequestHeader("X-User-Id") final Long memberId
     );
