+++ src/main/java/info/reinput/reinputworkspaceservice/folder/infra/client/ContentFeignAdapter.java
@@ -30,11 +30,9 @@ public class ContentFeignAdapter implements ContentPort {
     }
 
     @Override
-    public Integer copyInsight(final Long folderId, final Long memberId) {
+    public Long copyInsight(final Long folderId, final Long memberId) {
         log.info("copyInsight folderId : {}, memberId : {}", folderId, memberId);
 
         return contentClient.copyInsight(folderId, memberId).data();
     }
-
-
 }
